// src/app/shared/interfaces/person.ts
// export interface Person {
//     givenName: string;
//     surName: string;
//     age: number;
//     email: string;
//     address: {
//         area: string,
//         road: string
//     }
// }

// OR

// export interface Address {
//     area: string;
//     road: string
// }

// export interface Person {
//     givenName: string;
//     surName: string;
//     age: number;
//     email: string;
//     address: Address;
// }

// OR

export interface Person {
    givenName: string;
    surName: string;
    age: number;
    email: string;
    address: string
    }